+++ adapter/spark/src/main/java/com/asml/apa/wta/spark/driver/WtaDriverPlugin.java
@@ -5,16 +5,19 @@ import com.asml.apa.wta.core.config.RuntimeConfig;
 import com.asml.apa.wta.core.io.DiskOutputFile;
 import com.asml.apa.wta.core.io.OutputFile;
 import com.asml.apa.wta.core.model.Resource;
+import com.asml.apa.wta.core.model.ResourceState;
 import com.asml.apa.wta.core.model.Task;
 import com.asml.apa.wta.core.model.Workflow;
 import com.asml.apa.wta.core.model.Workload;
 import com.asml.apa.wta.spark.datasource.SparkDataSource;
+import com.asml.apa.wta.spark.dto.ResourceAndStateWrapper;
 import com.asml.apa.wta.spark.dto.ResourceCollectionDto;
 import com.asml.apa.wta.spark.streams.MetricStreamingEngine;
 import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.spark.SparkContext;
@@ -60,18 +63,19 @@ public class WtaDriverPlugin implements DriverPlugin {
     Map<String, String> executorVars = new HashMap<>();
     try {
       RuntimeConfig runtimeConfig = RuntimeConfig.readConfig();
+      this.metricStreamingEngine = new MetricStreamingEngine();
       sparkDataSource = new SparkDataSource(sparkCtx, runtimeConfig);
-      metricStreamingEngine = new MetricStreamingEngine();
       outputFile = new DiskOutputFile(Path.of(runtimeConfig.getOutputPath()));
       initListeners();
       executorVars.put("resourcePingInterval", String.valueOf(runtimeConfig.getResourcePingInterval()));
       executorVars.put(
           "executorSynchronizationInterval",
           String.valueOf(runtimeConfig.getExecutorSynchronizationInterval()));
+      executorVars.put("errorStatus", "false");
     } catch (Exception e) {
-      log.error(String.valueOf(e));
+      log.error("Error initialising WTA driver plugin, {} : {}.", e.getClass(), e.getMessage());
+      executorVars.put("errorStatus", "true");
       error = true;
-      shutdown();
     }
     return executorVars;
   }
@@ -106,28 +110,45 @@ public class WtaDriverPlugin implements DriverPlugin {
   @Override
   public void shutdown() {
     if (error) {
-      log.error("Error initialising WTA plugin. Shutting down plugin");
-    } else {
-      try {
-        removeListeners();
-        List<Task> tasks = sparkDataSource.getRuntimeConfig().isStageLevel()
-            ? sparkDataSource.getStageLevelListener().getProcessedObjects()
-            : sparkDataSource.getTaskLevelListener().getProcessedObjects();
-        List<Workflow> workflows = sparkDataSource.getJobLevelListener().getProcessedObjects();
-        List<Resource> resources = List.of();
-        Workload workload = sparkDataSource
-            .getApplicationLevelListener()
-            .getProcessedObjects()
-            .get(0);
-        WtaWriter wtaWriter = new WtaWriter(outputFile, "schema-1.0");
-        wtaWriter.write(Task.class, tasks);
-        wtaWriter.write(Resource.class, resources);
-        wtaWriter.write(Workflow.class, workflows);
-        wtaWriter.write(workload);
-      } catch (Exception e) {
-        log.error("Error while writing to the generated files, {} : {}.", e.getClass(), e.getMessage());
-      }
+      log.error("Plugin shutting down without generating files");
+      return;
     }
+    try {
+      endApplicationAndWrite();
+    } catch (Exception e) {
+      log.error("A {} error occurred while generating files: {}", e.getClass(), e.getMessage());
+    }
+  }
+
+  /**
+   * Removes the listeners and writes the collected data to the output file.
+   *
+   * @author Henry Page
+   * @since 1.0.0
+   */
+  private void endApplicationAndWrite() {
+    removeListeners();
+    List<Task> tasks = sparkDataSource.getRuntimeConfig().isStageLevel()
+        ? sparkDataSource.getStageLevelListener().getProcessedObjects()
+        : sparkDataSource.getTaskLevelListener().getProcessedObjects();
+    List<Workflow> workflows = sparkDataSource.getJobLevelListener().getProcessedObjects();
+    List<ResourceAndStateWrapper> resourceAndStateWrappers = metricStreamingEngine.collectResourceInformation();
+    List<Resource> resources = resourceAndStateWrappers.stream()
+        .map(ResourceAndStateWrapper::getResource)
+        .collect(Collectors.toList());
+    List<ResourceState> resourceStates = resourceAndStateWrappers.stream()
+        .flatMap(rs -> rs.getStates().stream())
+        .collect(Collectors.toList());
+    Workload workload = sparkDataSource
+        .getApplicationLevelListener()
+        .getProcessedObjects()
+        .get(0);
+    WtaWriter wtaWriter = new WtaWriter(outputFile, "schema-1.0");
+    wtaWriter.write(Task.class, tasks);
+    wtaWriter.write(Resource.class, resources);
+    wtaWriter.write(Workflow.class, workflows);
+    wtaWriter.write(ResourceState.class, resourceStates);
+    wtaWriter.write(workload);
   }
 
   /**
@@ -144,6 +165,7 @@ public class WtaDriverPlugin implements DriverPlugin {
     this.sparkDataSource.registerJobListener();
     this.sparkDataSource.registerApplicationListener();
   }
+
   /**
    * Removes the listeners.
    *