+++ adapter/spark/src/main/java/com/asml/apa/wta/spark/listener/JobLevelListener.java
@@ -64,7 +64,7 @@ public class JobLevelListener extends AbstractListener<Workflow> {
         .toArray(Task[]::new);
     final int numTasks = tasks.length;
     // we can also get the mode from the config, if that's what the user wants?
-    final String scheduler = "DAGScheduler";
+    final String scheduler = sparkContext.getConf().get("spark.scheduler.mode", "FIFO");
     final Domain domain = config.getDomain();
     final String appName = sparkContext.appName();