+++ core/src/main/java/com/asml/apa/wta/core/model/Workload.java
@@ -4,6 +4,7 @@ import com.asml.apa.wta.core.io.ParquetSchema;
 import com.asml.apa.wta.core.model.enums.Domain;
 import lombok.Builder;
 import lombok.Data;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.avro.generic.GenericRecord;
 
 /**
@@ -11,8 +12,10 @@ import org.apache.avro.generic.GenericRecord;
  *
  * @author Lohithsai Yadala Chanchu
  * @author Atour Mousavi Gourabi
+ * @author Tianchen Qu
  * @since 1.0.0
  */
+@Slf4j
 @Data
 @Builder
 public class Workload implements BaseTraceObject {
@@ -166,13 +169,15 @@ public class Workload implements BaseTraceObject {
   /**
    * This method should never be called as we do not need to output workloads in Parquet.
    *
-   * @param schema schema
+   * @param schema The parquet schema
    * @throws RuntimeException always
    * @author Atour Mousavi Gourabi
    * @since 1.0.0
    */
   @Override
   public final GenericRecord convertToRecord(ParquetSchema schema) {
+    log.error(
+        "The application attempted to convert a Workload to parquet, this is illegal and should never happen.");
     throw accessError();
   }
 
@@ -185,17 +190,7 @@ public class Workload implements BaseTraceObject {
    */
   @Override
   public final long getId() {
+    log.error("The application attempted to get the id of a Workload, this is illegal and should never happen.");
     throw accessError();
-  }
-
-  /**
-   * Creates a simple {@link RuntimeException}.
-   *
-   * @return a {@link RuntimeException} with a simple error message
-   * @author Atour Mousavi Gourabi
-   * @since 1.0.0
-   */
-  private RuntimeException accessError() {
-    return new RuntimeException("Something went wrong, this method shouldn't be called!");
   }
 }