+++ core/src/main/java/com/asml/apa/wta/core/supplier/IostatSupplier.java
@@ -5,6 +5,7 @@ import com.asml.apa.wta.core.utils.ShellUtils;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.stream.Collectors;
@@ -70,7 +71,7 @@ public class IostatSupplier implements InformationSupplier<IostatDto> {
    * @since 1.0.0
    */
   @Override
-  public CompletableFuture<IostatDto> getSnapshot() {
+  public CompletableFuture<Optional<IostatDto>> getSnapshot() {
     if (!this.isAvailable) {
       return notAvailableResult();
     }
@@ -78,25 +79,25 @@ public class IostatSupplier implements InformationSupplier<IostatDto> {
     CompletableFuture<String> allMetrics = shellUtils.executeCommand("iostat -d");
 
     return allMetrics.thenApply(result -> {
-      if (result != null) {
+      try {
         List<OutputLine> rows = parseIostat(result);
         double[] metrics = aggregateIostat(rows);
 
-        try {
-          return IostatDto.builder()
-              .tps(metrics[0])
-              .kiloByteReadPerSec(metrics[1])
-              .kiloByteWrtnPerSec(metrics[2])
-              .kiloByteDscdPerSec(metrics[3])
-              .kiloByteRead(metrics[4])
-              .kiloByteWrtn(metrics[5])
-              .kiloByteDscd(metrics[6])
-              .build();
-        } catch (Exception e) {
-          log.error("Something went wrong while receiving the iostat shell command outputs.");
-        }
+        return Optional.of(IostatDto.builder()
+            .tps(metrics[0])
+            .kiloByteReadPerSec(metrics[1])
+            .kiloByteWrtnPerSec(metrics[2])
+            .kiloByteDscdPerSec(metrics[3])
+            .kiloByteRead(metrics[4])
+            .kiloByteWrtn(metrics[5])
+            .kiloByteDscd(metrics[6])
+            .build());
+      } catch (NullPointerException npe) {
+        log.error("Iostat returned a malformed output: {}", npe.toString());
+      } catch (Exception e) {
+        log.error("Something went wrong while receiving the iostat bash command outputs.");
       }
-      return null;
+      return Optional.empty();
     });
   }